backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.133.0
h11==0.16.0
idna==3.11
numpy==2.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
PyJWT==2.12.1
resend==2.26.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
